thebase/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
httpx==0.27.0
jinja2==3.1.4
python-dotenv==1.0.1
email-validator==2.2.0
